C++11 mem::SharedPtr now inherits from std::shared_ptr
