+++ modules/c++/mem/unittests/test_shared_ptr.cpp
@@ -20,6 +20,7 @@
  *
  */
 
+#include <sys/sys_config.h>
 #include <mem/SharedPtr.h>
 
 #include "TestCase.h"
@@ -62,6 +63,13 @@ struct BarPtrTest : public FooPtrTest
 
     mem::SharedPtr<Bar> mBarPtr;
 };
+
+#ifdef __CODA_CPP11
+size_t cpp11Function(std::shared_ptr<Foo> foo)
+{
+    return foo->mVal;
+}
+#endif
 
 TEST_CASE(testNullCopying)
 {
@@ -249,6 +257,21 @@ TEST_CASE(testCasting)
         TEST_ASSERT_EQ(barPtr.getCount(), 3);
     }
 }
+
+#ifdef __CODA_CPP11
+TEST_CASE(testStdSharedPtr)
+{
+    const mem::SharedPtr<Foo> fooLegacy(new Foo(123));
+    std::shared_ptr<Foo> fooCtor(fooLegacy);
+    TEST_ASSERT_EQ(fooLegacy.get(), fooCtor.get());
+
+    std::shared_ptr<Foo> fooAssign = fooLegacy;
+    TEST_ASSERT_EQ(fooLegacy.get(), fooAssign.get());
+
+    TEST_ASSERT_EQ(cpp11Function(fooLegacy), 123);
+}
+#endif
+
 }
 
 int main(int, char**)
@@ -260,6 +283,9 @@ int main(int, char**)
    TEST_CHECK(testAssigning);
    TEST_CHECK(testSyntax);
    TEST_CHECK(testCasting);
+#ifdef __CODA_CPP11
+   TEST_CHECK(testStdSharedPtr);
+#endif
 
    return 0;
 }
